Automatic feeding buffer when not overloaded -------------------------------------------- Previous solution is expected as to calculate the accurate buffer size and feed rate, i.e., find out the duration and byte size of the audio file, feed size/duration bytes of data each epoch. This solution requires the audio file stream to know the duration of the audio, which is not always available.
The new solution watches carefully on the number of queued buffers in the audio queue.
When this number exceeds a given threshold, the timer's callback does nothing in that epoch.
The timer callback only feeds data when the audio queue is not overloaded with buffers
and the buffer size is fixed to 40000 bytes, which is approximately 1 second of audio data.
The timer interval is set to 0.5 second to prevent the starving queue situation.
--
	modified:   mypodcast/DZAudioPlayer.h
	modified:   mypodcast/DZAudioPlayer.mm
	modified:   mypodcast/DZAudioQueuePlayer.cpp
	modified:   mypodcast/DZAudioQueuePlayer.h